Do not dump stack traces of threads on test failure
Reverts 878301c795e6ad2c47c620c893b3324a564e0d91, relates #12425
