--- core/src/test/java/org/elasticsearch/test/StackTraces.java
@@ -1,63 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.test;
-
-import java.util.Formatter;
-import java.util.Locale;
-import java.util.Map;
-
-public class StackTraces {
-    /** Dump threads and their current stack trace. */
-    public static String formatThreadStacks(Map<Thread, StackTraceElement[]> threads) {
-        StringBuilder message = new StringBuilder();
-        int cnt = 1;
-        final Formatter f = new Formatter(message, Locale.ENGLISH);
-        for (Map.Entry<Thread, StackTraceElement[]> e : threads.entrySet()) {
-            if (e.getKey().isAlive()) {
-                f.format(Locale.ENGLISH, "\n  %2d) %s", cnt++, threadName(e.getKey())).flush();
-            }
-            if (e.getValue().length == 0) {
-                message.append("\n        at (empty stack)");
-            } else {
-                for (StackTraceElement ste : e.getValue()) {
-                    message.append("\n        at ").append(ste);
-                }
-            }
-        }
-        return message.toString();
-    }
-
-    private static String groupName(ThreadGroup threadGroup) {
-        if (threadGroup == null) {
-            return "{null group}";
-        } else {
-            return threadGroup.getName();
-        }
-    }
-
-    private static String threadName(Thread t) {
-        return "Thread[" +
-                "id=" + t.getId() +
-                ", name=" + t.getName() +
-                ", state=" + t.getState() +
-                ", group=" + groupName(t.getThreadGroup()) +
-                "]";
-    }
-}
